Fixes problem with detecting HTTP dump request via HTTP user
```
http://http-dump@localhost:8000/user/3/update
         ^^^
```

fixes #108

---

AI commit text

Updated also the test files and methods dealing with the handling of HTTP dumps, Inspector and Sentry events to include test cases that cover the scenario when an HTTP user is present. The goal of this change is to ensure handling even when an HTTP user is specified. In `AnyHttpRequestDump.php`, a condition was added to check if the HTTP user equals "http-dump", which is anticipated in the newly added scenarios. In the test files, new test methods were written to simulate requests with an HTTP user and additional assertions were made to verify if the events were correctly processed.